+++ app/src/main/assets/workout_en.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title> Workout</title>
+    <style>
+            body {background-color: #f1eeee;}
+            p {font-family: "Console", "Courier New", "Verdana";}
+            ol{font-family:"Console", "Courier New", "Verdana";}
+        </style>
+</head>
+<body>
+<p>If your goal is to lose weight and get fit, you must establish a weight training and cardio routine. By adding muscle with weight training, you increase your
+    metabolism and your ability to burn fat throughout the day. </p>
+<p>The following three routines were designed by <strong>Pat Manuel</strong> and serve as a good starting point to developing strength and overall fitness. The workout routines are all circuit style with little rest, so they will raise your heart level and keep you burning fat. Include 30 minutes of cardio on the days you are not weight training to see the best results.</p>
+<p>These exercises are to be performed once per week with a day of rest in-between each workout. The goal is to perform each circuit with as little rest as possible between each exercise in the circuit.</p>
+</body>
+</html>
