+++ app/src/main/java/com/hfad/freeingourselves/WorkoutIntroFragment.java
@@ -1,74 +1,53 @@
 package com.hfad.freeingourselves;
 
 
-import android.content.Context;
-import android.content.Intent;
-import android.database.Cursor;
-import android.database.sqlite.SQLiteException;
-import android.database.sqlite.SQLiteOpenHelper;
-import android.os.AsyncTask;
-import android.os.Bundle;
 import android.app.Fragment;
+import android.app.FragmentTransaction;
+import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.CursorAdapter;
-import android.widget.ListView;
-import android.widget.SimpleCursorAdapter;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import java.util.ArrayList;
+import android.webkit.WebView;
+import android.widget.Button;
 
 //intro to workouts and a list of workouts
 public class WorkoutIntroFragment extends Fragment {
 
     protected View view;
-    protected ListView listFavorites;
+
+    protected WebView wWebView;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
         view = inflater.inflate(R.layout.fragment_workout_intro, container, false);
 
-        // Populate the list_favorites ListView from a cursor.
-        listFavorites = (ListView) view.findViewById(R.id.workout_list);
+        wWebView = (WebView) view.findViewById(R.id.webView);
+        wWebView.loadUrl("file:///android_asset/workout_en.html");
 
-        new GetWorkoutNamesTask().execute(view.getContext());
-
-        // Navigate to WorkoutActivity if a drink is clicked.
-        listFavorites.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        Button startWorkoutButton = (Button) view.findViewById(R.id.startWorkoutButton);
+        startWorkoutButton.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onItemClick(AdapterView<?> listView, View v, int position, long id) {
-                Intent intent = new Intent(v.getContext(), WorkoutActivity.class);
-                intent.putExtra(WorkoutActivity.FAVE_NUM, (int) id);
-                startActivity(intent);
+            public void onClick(View view) {
+                switch (view.getId()) {
+                    case R.id.startWorkoutButton:
+                        Fragment workoutFragment = new WorkoutFragment();
+                        FragmentTransaction ft = getFragmentManager().beginTransaction();
+                        ft.replace(R.id.content_frame, workoutFragment);
+                        ft.addToBackStack(null);
+                        ft.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE);
+                        ft.commit();
+                        break;
+                }
             }
         });
 
         return view;
     }
 
-    private class GetWorkoutNamesTask extends AsyncTask<Context, Void, Cursor> {
-
-        protected Cursor doInBackground(Context... context) {
-            return FreeingOurselvesDatabaseUtilities.getWorkoutNames(context[0]);
-        }
-
-        protected void onPostExecute(Cursor workoutCursor) {
-            if (workoutCursor == null) {
-                Toast toast = Toast.makeText(view.getContext(), "Could not get workouts", Toast.LENGTH_SHORT);
-                toast.show();
-            } else {
-                CursorAdapter workoutAdapter = new SimpleCursorAdapter(view.getContext(),
-                        android.R.layout.simple_list_item_1, workoutCursor, new String[]{"NAME"},
-                        new int[]{android.R.id.text1}, 0);
-                listFavorites.setAdapter(workoutAdapter);
-            }
-        }
-    }
 }
+
+
+
+
