+++ app/src/main/res/layout/fragment_workout_intro.xml
@@ -1,21 +1,21 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    tools:context="com.hfad.freeingourselves.WorkoutIntroFragment">
-
-
-    <TextView
-        android:id="@+id/workout_intro"
-        android:layout_width="match_parent"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:tools="http://schemas.android.com/tools"
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+android:orientation="vertical"
+tools:context="com.hfad.freeingourselves.WorkoutIntroFragment">
+    <Button
+        android:text="@string/startWorkout"
+        android:id="@+id/startWorkoutButton"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/workout_intro" />
+        android:layout_gravity="center_horizontal"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"/>
 
-    <ListView
-        android:id="@+id/workout_list"
+    <WebView
+        android:id="@+id/webView"
+        android:layout_height="match_parent"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-
-</LinearLayout>
+        android:layout_above="@+id/startWorkoutButton"/>
+</RelativeLayout>
