+++ app/src/main/res/values/strings.xml
@@ -55,6 +55,7 @@
         and trans people of people of color; work that transforms the lives of women and girls;
         and introduces new alliances and tools for challenging racism, sexism, homophobia,
         and transphobia across our communities.</string>
+    <string name = "startWorkout">Start Workout</string>
 
 
     <string name="health_notes_button">Click here to go to saved questions</string>
